🐛 (android): fix configure_fingerprint.bin files ✨ (lib): implement subjects feature 🐛 (lib/core/constants/app_url.dart): fix subject url 🐛 (lib/core/constants/dummy_data.dart): fix semester type ✨ (lib/core/injection/src/subjects_injection.dart): change registerFactory to registerLazySingleton ✨ (lib/core/repo/auth_repo/repo_imp.dart): add number of major years to user model ✨ (lib/features/access/presentation/ui/screens/choose_access_method_screen.dart): remove user print ✨ (lib/features/subjects): implement subjects feature ✨ (lib/features/subjects/presentation/ui/screens/choose_years_screen.dart): get max year from user model ✨ (lib/shared/entities/user.dart): add number of major years to user model ✨ (lib/shared/services/dio_client.dart): fix refresh token logic ✨ (lib/shared/services/refresh_token_helper.dart): add number of major years to user model ✨ (lib/shared/widgets/custom_network_image.dart): fix error image
The configure_fingerprint.bin files were updated.
The subjects feature was implemented.
The subject url was fixed.
The semester type was fixed.
The registerFactory was changed to registerLazySingleton.
The number of major years was added to the user model.
The user print was removed.
The max year is now gotten from the user model.
The number of major years was added to the user model.
The refresh token logic was fixed.
The number of major years was added to the user model.
The error image was fixed.

+++ lib/features/subjects/domain/entities/subject_details_entity.dart
@@ -8,8 +8,8 @@ class SubjectDetailsEntity {
   final String subjectDescription;
   final String semester;
   final bool isPassed;
-  final int? midTermGrade;
-  final int? finalGrade;
+  final double? midTermGrade;
+  final double? finalGrade;
   final bool canEnroll;
   final bool doesHaveALab;
   final bool isMultipleChoice;
